The minimum set of changes to the "mbedtls" library required to establish a session with the Currant backend (see caveats!)
Most of the changes were required to enable support for the ecliptic
curve crytpo curves and ciphersuites the Currant backend likes to use
(or is choosing to use to make life easy for the ESP8266, since the
ECC stuff is less computation-intensive and doesn't require the xfer
of such large blobs of data).

One key change gleaned from message board posts from folks having
trouble getting "mbedtls" to work on the ESP8266 platform with their
particular servers was that MBEDTLS_SSL_MAX_CONTENT_LEN needs to be
set to (at least) 5120 bytes. My "test vehicle" for the mbedtls
library was the "esp8266-rtos-sample-code/04Protocal/Mbedtls_demo"
sample; depending on the resource requirements of our actual ESP8266
code we may need to tweak this up or down...

The final hurdle to clear in proving that this mbedtls implementation
would work with our backend was the need to adjust the server side
read timeout from the nominal 2 seconds to... 60 seconds!! That is how
amazingly slow the crypto numbers were crunched on the
ESP8266. Luckily we were testing against a clone of the actual Currant
backend service and were free to make these kind of changes...

Enabling the MBEDTLS_ECP_NIST_OPTIM configuration option vastly
improved our crypto-crunching speed: the time required perform the
SSL/TLS handshake dropped from around 60 seconds to 12 seconds. This
is still too slow for the actual backend service, but Hasty is going
to experiment with a couple more ecliptic curve implementations that
the ESP8266 might be able to process even faster.

Stay tuned.