fix: Tools for reformatting code.
This adds two tools: tools/reformat and tools/find-files. The first is
the main tool to be used by developers, the second is a utility used by
the first.

`tools/reformat` can replace tabs with spaces and run clang-format.

`tools/find-files` will list all files in the repository and tag each file
if it should have tabs replaced with spaced and if clang-format should
be run on it. The file patterns are configured in `tools/find-files.ini`

README.md is updated with some information on the code style and how to
run the reformat tool.