fix: Tools for reformatting code.
This adds two tools: tools/code-style and tools/tag-files. The first is
the main tool to be used by developers, the second is a utility used by
the first.

`tools/code-style` can replace tabs with spaces and run astyle.

`tools/tag-files` will list files and tag each file accoring to the
configuration file. The file patterns are configured in
`tools/code-style.ini`

README.md is updated with some information on the code style and how to
run the reformat tool.