utils: sha512: write only 48 bytes if hash type is SHA384
The esp_sha512_finish() function writes the calculated hash to the
dest buffer. Since this function is used for both SHA512 and SHA384
hash types, it should check which type is being used at a given
call, and write the corresponding number of bytes (either 48 or 64)
accordingly, otherwise if it unconditionally writes 64 bytes, when
called for an SHA384 hash it will write past the length of the
destination buffer.
This commit fixes the above issue by adding a new struct member to
esp_sha512_t which indicates whether SHA384 or SHA512 is being
used, and adding a check on the value of the struct member in
esp_sha512_finish() in order to write the correct number of bytes.