Don't override toolchain compiler variables if they are already set.
In case CMAKE_C_COMPILER and CMAKE_CXX_COMPILER variables are set,
we should not override them by default but rather check if they
point to a valid compiler. In case they do, leave them be.

Is is the use case of CLion IDE, when user defines a toolchain,
selects it in build configuration, that in turn generates
CMAKE_C_COMPILER and CMAKE_CXX_COMPILER variables basing on that,
but IDF scripts are overriding it by a hardcoded value.