Pass compiler settings to bootloader external project
Bootloader project is not getting CMAKE_C_COMPILER and CMAKE_CXX_COMPILER
variables from original CMake invocation but rather tries to figure it out
with defaults. Since those variables are being overwritten, we must also
pass them to any external projects that are being build as a part of a process.